feat: split cli and daemon
