+++ forest/cli/Cargo.toml
@@ -1,0 +1,93 @@
+[package]
+name        = "forest-cli"
+version     = "0.2.2"
+authors     = ["ChainSafe Systems <info@chainsafe.io>"]
+description = "Filecoin implementation in Rust. This crate contains all the subcommands except the daemon."
+edition     = "2021"
+
+[dependencies]
+anyhow.workspace                 = true
+async-log                        = "2.0"
+async-std                        = { workspace = true, features = ["attributes", "tokio1"] }
+base64.workspace                 = true
+byte-unit                        = "4.0"
+chrono                           = "0.4"
+cid                              = { workspace = true, default-features = false, features = ["std"] }
+directories                      = "4.0.1"
+fil_actor_miner_v8.workspace     = true
+forest_actor_interface.workspace = true
+forest_auth.workspace            = true
+forest_beacon.workspace          = true
+forest_blocks.workspace          = true
+forest_chain.workspace           = true
+forest_chain_sync.workspace      = true
+forest_cli_shared.path           = "../shared"
+forest_crypto.workspace          = true
+forest_db.workspace              = true
+forest_encoding.workspace        = true
+forest_fil_types.workspace       = true
+forest_interpreter.workspace     = true
+forest_ipld_blockstore.workspace = true
+forest_json.workspace            = true
+forest_key_management.workspace  = true
+forest_libp2p.workspace          = true
+forest_message                   = { workspace = true, features = ["blst"] }
+forest_message_pool.workspace    = true
+forest_net_utils.workspace       = true
+forest_networks.workspace        = true
+forest_paramfetch.workspace      = true
+forest_rpc-api.workspace         = true
+forest_rpc-client.workspace      = true
+forest_state_manager.workspace   = true
+forest_utils.workspace           = true
+forest_vm.workspace              = true
+futures.workspace                = true
+fvm.workspace                    = true
+fvm_ipld_car.workspace           = true
+fvm_ipld_encoding.workspace      = true
+fvm_shared                       = { workspace = true, default-features = false }
+git-version                      = "0.3"
+hex.workspace                    = true
+lazy_static.workspace            = true
+libp2p                           = { workspace = true, default-features = false, features = ["identify"] }
+log.workspace                    = true
+multibase.workspace              = true
+num-rational.workspace           = true
+num-traits.workspace             = true
+once_cell.workspace              = true
+pbr.workspace                    = true
+pretty_env_logger.workspace      = true
+prometheus                       = { workspace = true, features = ["process"] }
+rayon.workspace                  = true
+reqwest                          = { version = "0.11", features = ["stream"] }
+rpassword                        = "7.0"
+rug                              = "1.17"
+rust-s3                          = "0.32.3"
+serde                            = { workspace = true, features = ["derive"] }
+serde_json.workspace             = true
+sha2                             = "0.10.5"
+strfmt                           = "0.2.2"
+structopt                        = "0.3"
+tempfile                         = "3.3"
+ticker                           = "0.1"
+time                             = "0.3"
+tokio                            = { workspace = true, features = ["sync"] }
+toml                             = "0.5"
+uuid                             = { version = "1.1", features = ["v4"] }
+
+
+[dependencies.jsonrpc-v2]
+default-features = false
+features         = ["easy-errors", "macros", "bytes-v05"]
+version          = "0.11"
+
+[dev-dependencies]
+assert_cmd        = "2"
+quickcheck        = "1"
+quickcheck_macros = "1"
+rand.workspace    = true
+
+[features]
+default          = []
+testnet          = []
+release          = []
