+++ forest/daemon/Cargo.toml
@@ -1,0 +1,103 @@
+[package]
+name        = "forest-daemon"
+version     = "0.2.2"
+authors     = ["ChainSafe Systems <info@chainsafe.io>"]
+description = "Filecoin implementation in Rust. This command will start the daemon process."
+edition     = "2021"
+
+[[bin]]
+name = "forest"
+path = "src/main.rs"
+
+[dependencies]
+anyhow.workspace                 = true
+async-log                        = "2.0"
+async-std                        = { workspace = true, features = ["attributes", "tokio1"] }
+base64.workspace                 = true
+chrono                           = "0.4"
+cid                              = { workspace = true, default-features = false, features = ["std"] }
+ctrlc                            = "3.2"
+daemonize-me                     = "2.0"
+directories                      = "4.0.1"
+fil_actor_miner_v8.workspace     = true
+forest_actor_interface.workspace = true
+forest_auth.workspace            = true
+forest_beacon.workspace          = true
+forest_blocks.workspace          = true
+forest_chain.workspace           = true
+forest_chain_sync.workspace      = true
+forest_cli_shared.path           = "../shared"
+forest_crypto.workspace          = true
+forest_db.workspace              = true
+forest_deleg_cns                 = { workspace = true, optional = true }
+forest_encoding.workspace        = true
+forest_fil_cns                   = { workspace = true, optional = true }
+forest_fil_types.workspace       = true
+forest_genesis.workspace         = true
+forest_interpreter.workspace     = true
+forest_ipld_blockstore.workspace = true
+forest_json.workspace            = true
+forest_key_management.workspace  = true
+forest_libp2p.workspace          = true
+forest_message                   = { workspace = true, features = ["blst"] }
+forest_message_pool.workspace    = true
+forest_metrics.workspace         = true
+forest_net_utils.workspace       = true
+forest_networks.workspace        = true
+forest_paramfetch.workspace      = true
+forest_rpc-api.workspace         = true
+forest_rpc-client.workspace      = true
+forest_rpc.workspace             = true
+forest_state_manager.workspace   = true
+forest_utils.workspace           = true
+forest_vm.workspace              = true
+futures.workspace                = true
+fvm.workspace                    = true
+fvm_ipld_car.workspace           = true
+fvm_ipld_encoding.workspace      = true
+fvm_shared                       = { workspace = true, default-features = false }
+git-version                      = "0.3"
+hex.workspace                    = true
+lazy_static.workspace            = true
+libp2p                           = { workspace = true, default-features = false, features = ["identify"] }
+log.workspace                    = true
+multibase.workspace              = true
+num-rational.workspace           = true
+num-traits.workspace             = true
+once_cell.workspace              = true
+pbr.workspace                    = true
+pretty_env_logger.workspace      = true
+prometheus                       = { workspace = true, features = ["process"] }
+raw_sync                         = "0.1"
+rayon.workspace                  = true
+reqwest                          = { version = "0.11", features = ["stream"] }
+rpassword                        = "7.0"
+serde                            = { workspace = true, features = ["derive"] }
+serde_json.workspace             = true
+sha2                             = "0.10.5"
+shared_memory                    = "0.12"
+structopt                        = "0.3"
+tempfile                         = "3.3"
+time                             = "0.3"
+tokio                            = { workspace = true, features = ["sync"] }
+toml                             = "0.5"
+
+
+[dependencies.jsonrpc-v2]
+default-features = false
+features         = ["easy-errors", "macros", "bytes-v05"]
+version          = "0.11"
+
+[dev-dependencies]
+assert_cmd        = "2"
+quickcheck        = "1"
+quickcheck_macros = "1"
+rand.workspace    = true
+
+[features]
+default          = ["forest_fil_cns"]
+insecure_post    = ["forest_fil_cns/insecure_post"]
+testnet          = []
+release          = []
+forest_fil_cns   = ["dep:forest_fil_cns"]
+forest_deleg_cns = ["dep:forest_deleg_cns"]
