+++ forest/shared/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name        = "forest"
+name        = "forest_cli_shared"
 version     = "0.2.2"
 authors     = ["ChainSafe Systems <info@chainsafe.io>"]
-description = "Filecoin implementation in Rust. This command will start the daemon process."
+description = "Filecoin implementation in Rust. This crate contains shared utilities that power both cli and daemon."
 edition     = "2021"
 
 [dependencies]
@@ -10,15 +10,10 @@ anyhow.workspace                 = true
 async-log                        = "2.0"
 async-std                        = { workspace = true, features = ["attributes", "tokio1"] }
 base64.workspace                 = true
-byte-unit                        = "4.0"
-chrono                           = "0.4"
 cid                              = { workspace = true, default-features = false, features = ["std"] }
-ctrlc                            = "3.2"
-daemonize-me                     = "2.0"
 directories                      = "4.0.1"
 fil_actor_miner_v8.workspace     = true
 forest_actor_interface.workspace = true
-forest_auth.workspace            = true
 forest_beacon.workspace          = true
 forest_blocks.workspace          = true
 forest_chain.workspace           = true
@@ -27,9 +22,7 @@ forest_crypto.workspace          = true
 forest_db.workspace              = true
 forest_deleg_cns                 = { workspace = true, optional = true }
 forest_encoding.workspace        = true
-forest_fil_cns                   = { workspace = true, optional = true }
 forest_fil_types.workspace       = true
-forest_genesis.workspace         = true
 forest_interpreter.workspace     = true
 forest_ipld_blockstore.workspace = true
 forest_json.workspace            = true
@@ -37,13 +30,9 @@ forest_key_management.workspace  = true
 forest_libp2p.workspace          = true
 forest_message                   = { workspace = true, features = ["blst"] }
 forest_message_pool.workspace    = true
-forest_metrics.workspace         = true
-forest_net_utils.workspace       = true
 forest_networks.workspace        = true
-forest_paramfetch.workspace      = true
 forest_rpc-api.workspace         = true
 forest_rpc-client.workspace      = true
-forest_rpc.workspace             = true
 forest_state_manager.workspace   = true
 forest_utils.workspace           = true
 forest_vm.workspace              = true
@@ -64,24 +53,16 @@ once_cell.workspace              = true
 pbr.workspace                    = true
 pretty_env_logger.workspace      = true
 prometheus                       = { workspace = true, features = ["process"] }
-raw_sync                         = "0.1"
 rayon.workspace                  = true
 reqwest                          = { version = "0.11", features = ["stream"] }
-rpassword                        = "7.0"
-rug                              = "1.17"
-rust-s3                          = "0.32.3"
 serde                            = { workspace = true, features = ["derive"] }
 serde_json.workspace             = true
 sha2                             = "0.10.5"
-shared_memory                    = "0.12"
-strfmt                           = "0.2.2"
 structopt                        = "0.3"
 tempfile                         = "3.3"
-ticker                           = "0.1"
 time                             = "0.3"
 tokio                            = { workspace = true, features = ["sync"] }
 toml                             = "0.5"
-uuid                             = { version = "1.1", features = ["v4"] }
 
 
 [dependencies.jsonrpc-v2]
@@ -96,9 +77,3 @@ quickcheck_macros = "1"
 rand.workspace    = true
 
 [features]
-default          = ["forest_fil_cns"]
-insecure_post    = ["forest_fil_cns/insecure_post"]
-testnet          = []
-release          = []
-forest_fil_cns   = ["dep:forest_fil_cns"]
-forest_deleg_cns = ["dep:forest_deleg_cns"]
