+++ forest/shared/src/lib.rs
@@ -1,0 +1,5 @@
+// Copyright 2019-2022 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+pub mod cli;
+pub mod logger;
