--- forest/src/cli/mod.rs
@@ -1,498 +1,0 @@
-// Copyright 2019-2022 ChainSafe Systems
-// SPDX-License-Identifier: Apache-2.0, MIT
-
-mod auth_cmd;
-mod chain_cmd;
-mod client;
-mod config;
-mod config_cmd;
-mod fetch_params_cmd;
-mod genesis_cmd;
-mod mpool_cmd;
-mod net_cmd;
-mod snapshot_fetch;
-mod state_cmd;
-mod sync_cmd;
-mod wallet_cmd;
-
-pub(super) use self::auth_cmd::AuthCommands;
-pub(super) use self::chain_cmd::ChainCommands;
-pub use self::client::Client;
-pub use self::config::Config;
-pub use self::config::DaemonConfig;
-pub use self::config::LogConfig;
-pub use self::config::LogValue;
-pub(super) use self::fetch_params_cmd::FetchCommands;
-pub(super) use self::genesis_cmd::GenesisCommands;
-pub(super) use self::mpool_cmd::MpoolCommands;
-pub(super) use self::net_cmd::NetCommands;
-pub(super) use self::state_cmd::StateCommands;
-pub(super) use self::sync_cmd::SyncCommands;
-pub(super) use self::wallet_cmd::WalletCommands;
-
-use byte_unit::Byte;
-use directories::ProjectDirs;
-use forest_networks::ChainConfig;
-use futures::channel::oneshot::Receiver;
-use fvm_shared::bigint::BigInt;
-use fvm_shared::FILECOIN_PRECISION;
-use git_version::git_version;
-use jsonrpc_v2::Error as JsonRpcError;
-use log::{error, info, warn};
-use once_cell::sync::Lazy;
-use rug::float::ParseFloatError;
-use rug::Float;
-use serde::Serialize;
-use std::cell::RefCell;
-use std::io::{self, Write};
-use std::net::SocketAddr;
-use std::path::PathBuf;
-use std::process;
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::Arc;
-use structopt::StructOpt;
-
-use crate::cli::config_cmd::ConfigCommands;
-use cid::Cid;
-use forest_blocks::tipset_json::TipsetJson;
-use forest_utils::{read_file_to_string, read_toml};
-
-const GIT_HASH: &str = git_version!(args = ["--always", "--exclude", "*"], fallback = "unknown");
-
-pub static FOREST_VERSION_STRING: Lazy<String> =
-    Lazy::new(|| format!("{}+git.{}", env!("CARGO_PKG_VERSION"), GIT_HASH));
-
-/// CLI structure generated when interacting with Forest binary
-#[derive(StructOpt)]
-#[structopt(
-    name = env!("CARGO_PKG_NAME"),
-    version = FOREST_VERSION_STRING.as_str(),
-    about = env!("CARGO_PKG_DESCRIPTION"),
-    author = env!("CARGO_PKG_AUTHORS")
-)]
-pub struct Cli {
-    #[structopt(flatten)]
-    pub opts: CliOpts,
-    #[structopt(subcommand)]
-    pub cmd: Option<Subcommand>,
-}
-
-/// Forest binary sub-commands available.
-#[derive(StructOpt)]
-#[structopt(setting = structopt::clap::AppSettings::VersionlessSubcommands)]
-pub enum Subcommand {
-    /// Download parameters for generating and verifying proofs for given size
-    #[structopt(name = "fetch-params")]
-    Fetch(FetchCommands),
-
-    /// Interact with Filecoin blockchain
-    #[structopt(name = "chain")]
-    Chain(ChainCommands),
-
-    /// Manage RPC permissions
-    #[structopt(name = "auth")]
-    Auth(AuthCommands),
-
-    /// Work with blockchain genesis
-    #[structopt(name = "genesis")]
-    Genesis(GenesisCommands),
-
-    /// Manage P2P network
-    #[structopt(name = "net")]
-    Net(NetCommands),
-
-    /// Manage wallet
-    #[structopt(name = "wallet")]
-    Wallet(WalletCommands),
-
-    /// Inspect or interact with the chain synchronizer
-    #[structopt(name = "sync")]
-    Sync(SyncCommands),
-
-    /// Interact with the message pool
-    #[structopt(name = "mpool")]
-    Mpool(MpoolCommands),
-
-    /// Interact with and query Filecoin chain state
-    #[structopt(name = "state")]
-    State(StateCommands),
-
-    /// Manage node configuration
-    #[structopt(name = "config")]
-    Config(ConfigCommands),
-}
-
-/// CLI options
-#[derive(StructOpt, Debug)]
-pub struct CliOpts {
-    /// A TOML file containing relevant configurations
-    #[structopt(short, long)]
-    pub config: Option<String>,
-    /// The genesis CAR file
-    #[structopt(short, long)]
-    pub genesis: Option<String>,
-    /// Allow RPC to be active or not (default: true)
-    #[structopt(short, long)]
-    pub rpc: Option<bool>,
-    /// Client JWT token to use for JSON-RPC authentication
-    #[structopt(short, long)]
-    pub token: Option<String>,
-    /// Address used for metrics collection server. By defaults binds on localhost on port 6116.
-    #[structopt(long)]
-    pub metrics_address: Option<SocketAddr>,
-    /// Address used for RPC. By defaults binds on localhost on port 1234.
-    #[structopt(long)]
-    pub rpc_address: Option<SocketAddr>,
-    /// Allow Kademlia (default: true)
-    #[structopt(short, long)]
-    pub kademlia: Option<bool>,
-    /// Allow MDNS (default: false)
-    #[structopt(long)]
-    pub mdns: Option<bool>,
-    /// Validate snapshot at given EPOCH
-    #[structopt(long)]
-    pub height: Option<i64>,
-    /// Import a snapshot from a local CAR file or URL
-    #[structopt(long)]
-    pub import_snapshot: Option<String>,
-    /// Halt with exit code 0 after successfully importing a snapshot
-    #[structopt(long)]
-    pub halt_after_import: bool,
-    /// Import a chain from a local CAR file or URL
-    #[structopt(long)]
-    pub import_chain: Option<String>,
-    /// Skips loading CAR file and uses header to index chain. Assumes a pre-loaded database
-    #[structopt(long)]
-    pub skip_load: bool,
-    /// Number of tipsets requested over chain exchange (default is 200)
-    #[structopt(long)]
-    pub req_window: Option<i64>,
-    /// Number of tipsets to include in the sample that determines what the network head is
-    #[structopt(long)]
-    pub tipset_sample_size: Option<u8>,
-    /// Amount of Peers we want to be connected to (default is 75)
-    #[structopt(long)]
-    pub target_peer_count: Option<u32>,
-    /// Encrypt the key-store (default: true)
-    #[structopt(long)]
-    pub encrypt_keystore: Option<bool>,
-    /// Choose network chain to sync to
-    #[structopt(
-        long,
-        default_value = "mainnet",
-        possible_values = &["mainnet", "calibnet"],
-    )]
-    pub chain: String,
-    /// Daemonize Forest process
-    #[structopt(long)]
-    pub detach: bool,
-    // env_logger-0.7 can only redirect to stderr or stdout. Version 0.9 can redirect to a file.
-    // However, we cannot upgrade to version 0.9 because pretty_env_logger depends on version 0.7
-    // and hasn't been updated in quite a while. See https://github.com/seanmonstar/pretty-env-logger/issues/52
-    // #[structopt(
-    //     help = "Specify a filename into which logging should be appended"
-    // )]
-    // pub log_file: Option<PathBuf>,
-}
-
-impl CliOpts {
-    pub fn to_config(&self) -> Result<Config, io::Error> {
-        let mut cfg: Config = match &self.config {
-            Some(config_file) => {
-                // Read from config file
-                let toml = read_file_to_string(&PathBuf::from(&config_file))?;
-                // Parse and return the configuration file
-                read_toml(&toml)?
-            }
-            None => find_default_config().unwrap_or_default(),
-        };
-
-        if self.chain == "calibnet" {
-            // override the chain configuration
-            cfg.chain = Arc::new(ChainConfig::calibnet());
-        }
-
-        if let Some(genesis_file) = &self.genesis {
-            cfg.client.genesis_file = Some(genesis_file.to_owned());
-        }
-        if self.rpc.unwrap_or(cfg.client.enable_rpc) {
-            cfg.client.enable_rpc = true;
-            if let Some(rpc_address) = self.rpc_address {
-                cfg.client.rpc_address = rpc_address;
-            }
-
-            if self.token.is_some() {
-                cfg.client.rpc_token = self.token.to_owned();
-            }
-        } else {
-            cfg.client.enable_rpc = false;
-        }
-        if let Some(metrics_address) = self.metrics_address {
-            cfg.client.metrics_address = metrics_address;
-        }
-        if self.import_snapshot.is_some() && self.import_chain.is_some() {
-            cli_error_and_die(
-                "Can't set import_snapshot and import_chain at the same time!",
-                1,
-            );
-        } else {
-            if let Some(snapshot_path) = &self.import_snapshot {
-                cfg.client.snapshot_path = Some(snapshot_path.to_owned());
-                cfg.client.snapshot = true;
-            }
-            if let Some(snapshot_path) = &self.import_chain {
-                cfg.client.snapshot_path = Some(snapshot_path.to_owned());
-                cfg.client.snapshot = false;
-            }
-            cfg.client.snapshot_height = self.height;
-
-            cfg.client.skip_load = self.skip_load;
-        }
-
-        cfg.client.halt_after_import = self.halt_after_import;
-
-        cfg.network.kademlia = self.kademlia.unwrap_or(cfg.network.kademlia);
-        cfg.network.mdns = self.mdns.unwrap_or(cfg.network.mdns);
-        if let Some(target_peer_count) = self.target_peer_count {
-            cfg.network.target_peer_count = target_peer_count;
-        }
-        // (where to find these flags, should be easy to do with structops)
-
-        // check and set syncing configurations
-        // TODO add MAX conditions
-        if let Some(req_window) = &self.req_window {
-            cfg.sync.req_window = req_window.to_owned();
-        }
-        if let Some(tipset_sample_size) = self.tipset_sample_size {
-            cfg.sync.tipset_sample_size = tipset_sample_size.into();
-        }
-        if let Some(encrypt_keystore) = self.encrypt_keystore {
-            cfg.client.encrypt_keystore = encrypt_keystore;
-        }
-
-        Ok(cfg)
-    }
-}
-
-fn find_default_config() -> Option<Config> {
-    if let Ok(config_file) = std::env::var("FOREST_CONFIG_PATH") {
-        info!(
-            "FOREST_CONFIG_PATH detected, using configuration at {}",
-            config_file
-        );
-        let path = PathBuf::from(config_file);
-        if path.exists() {
-            return read_config_or_none(path);
-        }
-    };
-
-    if let Some(dir) = ProjectDirs::from("com", "ChainSafe", "Forest") {
-        let mut config_dir = dir.config_dir().to_path_buf();
-        config_dir.push("config.toml");
-        if config_dir.exists() {
-            info!("Found a config file at {}", config_dir.display());
-            return read_config_or_none(config_dir);
-        }
-    }
-
-    warn!("No configurations found, using defaults.");
-
-    None
-}
-
-fn read_config_or_none(path: PathBuf) -> Option<Config> {
-    let toml = match read_file_to_string(&path) {
-        Ok(t) => t,
-        Err(e) => {
-            warn!("An error occured while reading configuration file at {}. Resorting to default configuration. Error was {}", path.display(), e);
-            return None;
-        }
-    };
-
-    match read_toml(&toml) {
-        Ok(cfg) => Some(cfg),
-        Err(e) => {
-            warn!(
-                "Error reading configuration file, opting to defaults. Error was {} ",
-                e
-            );
-            None
-        }
-    }
-}
-
-pub fn set_sigint_handler() -> Receiver<()> {
-    let (ctrlc_send, ctrlc_oneshot) = futures::channel::oneshot::channel();
-    let ctrlc_send_c = RefCell::new(Some(ctrlc_send));
-
-    let running = Arc::new(AtomicUsize::new(0));
-    ctrlc::set_handler(move || {
-        let prev = running.fetch_add(1, Ordering::SeqCst);
-        if prev == 0 {
-            warn!("Got interrupt, shutting down...");
-            // Send sig int in channel to blocking task
-            if let Some(ctrlc_send) = ctrlc_send_c.try_borrow_mut().unwrap().take() {
-                ctrlc_send.send(()).expect("Error sending ctrl-c message");
-            }
-        } else {
-            info!("Exiting process");
-            process::exit(0);
-        }
-    })
-    .expect("Error setting Ctrl-C handler");
-
-    ctrlc_oneshot
-}
-
-/// Pretty-print a JSON-RPC error and exit
-pub(super) fn handle_rpc_err(e: JsonRpcError) -> ! {
-    match e {
-        JsonRpcError::Full {
-            code,
-            message,
-            data: _,
-        } => {
-            error!("JSON RPC Error: Code: {} Message: {}", code, message);
-            process::exit(code as i32);
-        }
-        JsonRpcError::Provided { code, message } => {
-            error!("JSON RPC Error: Code: {} Message: {}", code, message);
-            process::exit(code as i32);
-        }
-    }
-}
-
-/// Format a vector to a prettified string
-pub(super) fn format_vec_pretty(vec: Vec<String>) -> String {
-    format!("[{}]", vec.join(", "))
-}
-
-/// convert `BigInt` to size string using byte size units (i.e. KiB, GiB, PiB, etc)
-/// Provided number cannot be negative, otherwise the function will panic.
-pub(super) fn to_size_string(input: &BigInt) -> anyhow::Result<String> {
-    let bytes = u128::try_from(input)
-        .map_err(|e| anyhow::anyhow!("error parsing the input {}: {}", input, e))?;
-
-    Ok(Byte::from_bytes(bytes)
-        .get_appropriate_unit(true)
-        .to_string())
-}
-
-/// Print an error message and exit the program with an error code
-/// Used for handling high level errors such as invalid parameters
-pub(super) fn cli_error_and_die(msg: impl AsRef<str>, code: i32) -> ! {
-    error!("Error: {}", msg.as_ref());
-    std::process::exit(code);
-}
-
-/// Prints a plain HTTP JSON-RPC response result
-pub(super) fn print_rpc_res(res: Result<String, JsonRpcError>) {
-    match res {
-        Ok(obj) => println!("{}", &obj),
-        Err(err) => handle_rpc_err(err),
-    };
-}
-
-/// Prints a pretty HTTP JSON-RPC response result
-pub(super) fn print_rpc_res_pretty<T: Serialize>(res: Result<T, JsonRpcError>) {
-    match res {
-        Ok(obj) => println!("{}", serde_json::to_string_pretty(&obj).unwrap()),
-        Err(err) => handle_rpc_err(err),
-    };
-}
-
-/// Prints a tipset from a HTTP JSON-RPC response result
-pub(super) fn print_rpc_res_cids(res: Result<TipsetJson, JsonRpcError>) {
-    match res {
-        Ok(tipset) => println!(
-            "{}",
-            serde_json::to_string_pretty(
-                &tipset
-                    .0
-                    .cids()
-                    .iter()
-                    .map(|cid: &Cid| cid.to_string())
-                    .collect::<Vec<_>>()
-            )
-            .unwrap()
-        ),
-        Err(err) => handle_rpc_err(err),
-    };
-}
-
-/// Prints a bytes HTTP JSON-RPC response result
-pub(super) fn print_rpc_res_bytes(res: Result<Vec<u8>, JsonRpcError>) {
-    match res {
-        Ok(obj) => println!(
-            "{}",
-            String::from_utf8(obj)
-                .map_err(|e| handle_rpc_err(e.into()))
-                .unwrap()
-        ),
-        Err(err) => handle_rpc_err(err),
-    };
-}
-
-/// Prints a string HTTP JSON-RPC response result to a buffered `stdout`
-pub(super) fn print_stdout(out: String) {
-    let stdout = io::stdout();
-    let mut handle = stdout.lock();
-    handle
-        .write_all(out.as_bytes())
-        .map_err(|e| handle_rpc_err(e.into()))
-        .unwrap();
-
-    handle
-        .write("\n".as_bytes())
-        .map_err(|e| handle_rpc_err(e.into()))
-        .unwrap();
-}
-
-/// Convert an `attoFIL` balance to `FIL`
-pub(super) fn balance_to_fil(balance: BigInt) -> Result<Float, ParseFloatError> {
-    let raw = Float::parse_radix(balance.to_string(), 10)?;
-    let b = Float::with_val(128, raw);
-
-    let raw = Float::parse_radix(FILECOIN_PRECISION.to_string().as_bytes(), 10)?;
-    let p = Float::with_val(64, raw);
-
-    Ok(Float::with_val(128, b / p))
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use fvm_shared::bigint::Zero;
-
-    #[test]
-    fn to_size_string_valid_input() {
-        let cases = [
-            (BigInt::zero(), "0 B"),
-            (BigInt::from(1 << 10), "1024 B"),
-            (BigInt::from((1 << 10) + 1), "1.00 KiB"),
-            (BigInt::from((1 << 10) + 512), "1.50 KiB"),
-            (BigInt::from(1 << 20), "1024.00 KiB"),
-            (BigInt::from((1 << 20) + 1), "1.00 MiB"),
-            (BigInt::from(1 << 29), "512.00 MiB"),
-            (BigInt::from((1 << 30) + 1), "1.00 GiB"),
-            (BigInt::from((1u64 << 40) + 1), "1.00 TiB"),
-            (BigInt::from((1u64 << 50) + 1), "1.00 PiB"),
-            // ZiB is 2^70, 288230376151711744 is 2^58
-            (BigInt::from(u128::MAX), "288230376151711744.00 ZiB"),
-        ];
-
-        for (input, expected) in cases {
-            assert_eq!(to_size_string(&input).unwrap(), expected.to_string());
-        }
-    }
-
-    #[test]
-    fn to_size_string_negative_input_should_fail() {
-        assert!(to_size_string(&BigInt::from(-1i8)).is_err());
-    }
-
-    #[test]
-    fn to_size_string_too_large_input_should_fail() {
-        assert!(to_size_string(&(BigInt::from(u128::MAX) + 1)).is_err());
-    }
-}
